requirements.txt
colorama==0.4.4
docx2pdf==0.1.7
lxml==4.6.3
numpy==1.21.2
pandas==1.3.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
pywin32==227
six==1.16.0
tqdm==4.62.2
